// src/app/feature-modules/main/sidebarconfigure/tablerouting/data/constant-value.enum.ts
export enum Constant {
    CONTROLLER = "/SymTableRouting",
    PATH = "/main/configure/table-routing",
    UNCLICKABLE = "unclickable",
    CLICKABLE = "clickable",
    CREATE = "/create",
    EDIT = "/edit",
    DELETE = "/delete",
    SORT_BY = "trigger_id",
    SORT_ORDER = "asc",
    SAVE = "Save",
    UPDATE = "Update"
}